Create an exceptionhandler using generic handling with OpenAPI
Exception Handling with REST always has a common problem.
Somehow the exception needs to be mapped to an REST exception class.

This often includes the status code of an exception, as well as further
information.

A naiv approach would be to add those details into the ExceptionClass,
but this has two significant downsides. First of all the Rest specific
information like the HTTP Status Code do not belong into the exception.
They're part of the service layer.
Second system specific exceptions do not contain those information.

This approach here defines a generic mapper that can be controlled per
exception class and that can use a behaviour per information on how to
get the information.
The mapper is part of the service layer and can easily be adjusted.