Change ExceptionMapper to use definition Functions instead of setter
The Setters in the exceptionMappers directly put values into the
ProblemDetailsTo.
This can be considered a code smell, because the ProblemDetailsTo is
changed hidden as a reference. Therefore, now Functions are used, that
return the value that should be added to the ProblemDetailsTo.
The map method created the ProblemDetailsTo and adds the values from the
Functions.